Interactive dragging of image implemented
This required severe rework of internal data structures
to allow conversion of on-screen versus stage versus math coordinates.
Also, the CompStorage is now a first-class citizen and not hidden
in the computation engine any more as it has to survive computation
to be available for shifting.

I also started cleaning up representation of computation data.
Storing them as the strings of the input fields was never a good idea.
Now, this is at least for the stage math coodrinates replaced with
actual numerical storage as f64 values.

I also switched to the 2024 edition of Rust which, together with Rust
1.88, allows for this nifty "if let … && let …" syntax.