First working implementation of zooming into and out of the image
Use mouse wheel to zoom in and out of the point where the mouse
pointer sits on. Zoom state needed to go out of the canvas' state
into the application's state due to the application-controlled
timeout for the zoom operation.

Note that these changes have been implemented without AI support.