Implemented issue 4.15: Parallel dot compilation.
Simply amazing usage of the "rayon" library made this almost a one-liner.
This step has been done with the help of Claude.
Note that Claude did not really understand the stuff with the counter
in the computation loop. I added some optimisations here.

With the changes of issues 4.14 and 4.15 computation has become really
fast already.