Move StageProperties and ImageCompProperties from f64 to BigDecimal for number representation
This is a big one! To be able to perform arbitrary zooms,
f64 is not sufficient. Therefore, BigDecimal will be used if needed.
For this to work, BigDecimal must be used for all data representation
before it can be used in the actual point computation loop.

This is one of the main steps of this transition: StageProperties is one
of the most important framework representation and computation classes.
Internally, it uses (Rastered)MathArea for data representation and is
therefore completely converted into BigDecimal number representation.

Some f64-based API methods remain for the moment, but the general target
is to make BigDecimal the main representation class for numbers
in the complete program.